refactor(operator_commands_dashboard): extract distributed/vacate_vm + helpers
Move distributed and vacate_vm HTTP handlers plus remote_vms_from_hosts
and strip_ansi_codes helpers from routes.rs into a new sibling
distributed.rs module. Promote save_hosts to pub(crate).

- routes.rs: 4110 -> 3773
- distributed.rs (NEW, 346)

Refs #1266
